Refactor requests reducer and parsing to Typescript
