+++ cardstack/src/hooks/transaction-confirmation/use-confirm-transaction.ts
@@ -25,7 +25,7 @@ import {
 } from '@rainbow-me/utils/signingMethods';
 import logger from 'logger';
 import { walletConnectSendStatus } from '@rainbow-me/redux/walletconnect';
-import { removeRequest } from '@rainbow-me/redux/requests';
+import { removeRequest } from '@cardstack/redux/requests';
 import { dataAddNewTransaction } from '@rainbow-me/redux/data';
 import { toHex, estimateGasWithPadding } from '@rainbow-me/handlers/web3';
 import { useGas } from '@rainbow-me/hooks';
