+++ cardstack/src/redux/requests.ts
@@ -1,4 +1,6 @@
 import { filter, get, omit, values } from 'lodash';
+import { AnyAction } from '@reduxjs/toolkit';
+import { getRequestDisplayDetails } from '@cardstack/parsers/signing-requests';
 import {
   getLocalRequests,
   removeLocalRequest,
@@ -8,50 +10,61 @@ import {
   dappLogoOverride,
   dappNameOverride,
 } from '@rainbow-me/helpers/dappNameHandler';
-import { getRequestDisplayDetails } from '@rainbow-me/parsers';
 import logger from 'logger';
+import { AppDispatch, AppGetState } from '@rainbow-me/redux/store';
 
 // -- Constants --------------------------------------- //
 const REQUESTS_UPDATE_REQUESTS_TO_APPROVE =
   'requests/REQUESTS_UPDATE_REQUESTS_TO_APPROVE';
+
 const REQUESTS_CLEAR_STATE = 'requests/REQUESTS_CLEAR_STATE';
 
 // Requests expire automatically after 1 hour
 const EXPIRATION_THRESHOLD_IN_MS = 1000 * 60 * 60;
 
-export const requestsLoadState = () => async (dispatch, getState) => {
+export const requestsLoadState = () => async (
+  dispatch: AppDispatch,
+  getState: AppGetState
+) => {
   const { accountAddress, network } = getState().settings;
+
   try {
     const requests = await getLocalRequests(accountAddress, network);
     const _requests = requests || {};
     dispatch({ payload: _requests, type: REQUESTS_UPDATE_REQUESTS_TO_APPROVE });
-    // eslint-disable-next-line no-empty
   } catch (error) {}
 };
 
 export const addRequestToApprove = (
-  clientId,
-  peerId,
-  requestId,
-  payload,
-  peerMeta
-) => (dispatch, getState) => {
+  clientId: any,
+  peerId: any,
+  requestId: any,
+  payload: any,
+  peerMeta: any
+) => (dispatch: AppDispatch, getState: AppGetState) => {
   const { requests } = getState().requests;
   const { accountAddress, network, nativeCurrency } = getState().settings;
   const { assets } = getState().data;
+
   const displayDetails = getRequestDisplayDetails(
     payload,
     assets,
     nativeCurrency
   );
+
   const oneHourAgoTs = Date.now() - EXPIRATION_THRESHOLD_IN_MS;
+
   if (displayDetails.timestampInMs < oneHourAgoTs) {
     logger.log('request expired!');
+
     return;
   }
+
   const imageUrl = dappLogoOverride(peerMeta.url) || get(peerMeta, 'icons[0]');
+
   const dappName =
     dappNameOverride(peerMeta.url) || peerMeta.name || 'Unknown Dapp';
+
   const dappUrl = peerMeta.url || 'Unknown Url';
   const dappScheme = peerMeta.scheme || null;
 
@@ -66,24 +79,33 @@ export const addRequestToApprove = (
     peerId,
     requestId,
   };
+
   const updatedRequests = { ...requests, [requestId]: request };
   dispatch({
     payload: updatedRequests,
     type: REQUESTS_UPDATE_REQUESTS_TO_APPROVE,
   });
+
   saveLocalRequests(updatedRequests, accountAddress, network);
+
   return request;
 };
 
-export const requestsForTopic = topic => (dispatch, getState) => {
+export const requestsForTopic = (topic: any) => (
+  dispatch: AppDispatch,
+  getState: AppGetState
+) => {
   const { requests } = getState().requests;
   return filter(values(requests), { clientId: topic });
 };
 
-export const requestsResetState = () => dispatch =>
+export const requestsResetState = () => (dispatch: AppDispatch) =>
   dispatch({ type: REQUESTS_CLEAR_STATE });
 
-export const removeRequest = requestId => (dispatch, getState) => {
+export const removeRequest = (requestId: any) => (
+  dispatch: AppDispatch,
+  getState: AppGetState
+) => {
   const { accountAddress, network } = getState().settings;
   const { requests } = getState().requests;
   const updatedRequests = omit(requests, [requestId]);
@@ -99,7 +121,7 @@ const INITIAL_STATE = {
   requests: {},
 };
 
-export default (state = INITIAL_STATE, action) => {
+export default (state = INITIAL_STATE, action: AnyAction) => {
   switch (action.type) {
     case REQUESTS_UPDATE_REQUESTS_TO_APPROVE:
       return {
